% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptionDataSet-class.R
\docType{data}
\name{annot}
\alias{annot}
\title{Reference annotation (knownGene from UCSC)}
\format{
\code{\link[GenomicRanges:GRanges-class]{GRanges}} object
}
\usage{
annot
}
\description{
\code{annot} is an object of \code{\link[GenomicRanges:GRanges-class]{GRanges}}
class, containing genomic coordinates of the genes located on human
chromosome 15 (chr15:63261757-84081194).
}
\examples{
data(annot)
annot

}
\keyword{datasets}
