% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_transcript.R
\name{combine_jhu_transcript}
\alias{combine_jhu_transcript}
\alias{combine_transcript}
\title{Combine JHU Transcript}
\usage{
combine_jhu_transcript(file)

combine_transcript(file, type = "jhu", ...)
}
\arguments{
\item{file}{a vector of paths to PDFs}

\item{type}{Type of Transcript}

\item{...}{additional arguments to pass to combining function}
}
\value{
A \code{data_frame} of the data
}
\description{
Combine JHU Transcript
}
\examples{
file = "~/Downloads/Transcript.pdf"
if (file.exists(file)) {
    res = combine_jhu_transcript(c(file, file))
}
}
