% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipDataSet-generics.R, R/ChipDataSet-methods.R
\docType{methods}
\name{getConfusionMatrix}
\alias{getConfusionMatrix}
\alias{getConfusionMatrix,ChipDataSet-method}
\title{getConfusionMatrix}
\usage{
getConfusionMatrix(object)

\S4method{getConfusionMatrix}{ChipDataSet}(object)
}
\arguments{
\item{object}{A \code{\link{ChipDataSet}} object.}
}
\value{
An object of \code{Confusion Matrix} class. For the details see
    \href{http://topepo.github.io/caret/index.html}{caret} package.
}
\description{
Retrieve a cross-tabulation of observed and predicted classes (prediction of
gene associated peaks) with associated statistics.
}
\examples{
### Load ChipDataSet object
data(cds)

getConfusionMatrix(cds)

}
\seealso{
\code{\link{predictTssOverlap}} \code{\link[caret]{confusionMatrix}}
}
\author{
Armen R. Karapetyan
}
