% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipDataSet-generics.R, R/ChipDataSet-methods.R
\docType{methods}
\name{getPredictorSignificance}
\alias{getPredictorSignificance}
\alias{getPredictorSignificance,ChipDataSet-method}
\title{getPredictorSignificance}
\usage{
getPredictorSignificance(object)

\S4method{getPredictorSignificance}{ChipDataSet}(object)
}
\arguments{
\item{object}{A \code{\link{ChipDataSet}} object.}
}
\value{
A vector of p-values.
}
\description{
Retrieve significance of each predictor used in the classification model fit
(prediction of gene associated peaks).
}
\examples{
### Load ChipDataSet object
data(cds)

getPredictorSignificance(cds)

}
\seealso{
\code{\link{predictTssOverlap}}
}
\author{
Armen R. Karapetyan
}
