% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptionDataSet-class.R
\docType{data}
\name{tds}
\alias{tds}
\title{Example of \code{TranscriptionDataSet} object.}
\format{
\code{\link{TranscriptionDataSet}} object
}
\usage{
tds
}
\description{
\code{tds} is an object of \code{\link{TranscriptionDataSet}} class, containing
nuclear RNA-seq data for human chromosome 15 (chr15:63261757-84081194),
profiled in prostate cancer LNCaP cells.
}
\examples{
data(tds)
tds

}
\keyword{datasets}
