% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptr_shiny_example.R
\name{transcript_shiny_example}
\alias{transcript_shiny_example}
\title{Transcriptr Shiny Examples}
\usage{
transcript_shiny_example(example = c("validate_curriculum"))
}
\arguments{
\item{example}{Name of example to run.}
}
\value{
The output from \code{\link{runApp}}
}
\description{
Transcriptr Shiny Examples
}
\examples{
if (interactive()) {
transcript_shiny_example()
}
}
