% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptionDataSet-class.R
\docType{data}
\name{annot}
\alias{annot}
\title{Reference annotation (knownGene from UCSC)}
\format{\code{\link{GRanges}} object}
\usage{
annot
}
\value{
\code{\link{GRanges}} object
}
\description{
\code{annot} is an object of \code{\link{GRanges}} class, containing
genomic coordinates of the genes located on human chromosome 15 (chr15:63261757-84081194).
}
\keyword{datasets}

