% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipDataSet-class.R
\docType{data}
\name{cds}
\alias{cds}
\title{Example of \code{ChipDataSet} object.}
\format{\code{\link{ChipDataSet}} object}
\usage{
cds
}
\value{
\code{\link{ChipDataSet}} object
}
\description{
\code{cds} is an object of \code{\link{ChipDataSet}} class, containing
H3K4me3 active histone mark ChIP-seq peaks from human chromosome 15
(chr15:63261757-84081194) profiled in prostate cancer LNCaP cells.
}
\keyword{datasets}

