% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipDataSet-generics.R, R/ChipDataSet-methods.R
\docType{methods}
\name{getPeaks}
\alias{getPeaks}
\alias{getPeaks,ChipDataSet-method}
\title{getPeaks}
\usage{
getPeaks(object)

\S4method{getPeaks}{ChipDataSet}(object)
}
\arguments{
\item{object}{A \code{\link{ChipDataSet}} object.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object.
}
\description{
Retrieve ChIP-seq peak information from the \code{\link{ChipDataSet}} object.
}
\examples{
### Load ChipDataSet object
data(cds)

getPeaks(cds)

}
\author{
Armen R. Karapetyan
}
\seealso{
\code{\link{constructCDS}}
    \code{\link{predictTssOverlap}}
    \code{\link{predictStrand}}
}

