% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChipDataSet-generics.R, R/ChipDataSet-methods.R
\docType{methods}
\name{getProbTreshold}
\alias{getProbTreshold}
\alias{getProbTreshold,ChipDataSet-method}
\title{getProbTreshold}
\usage{
getProbTreshold(object)

\S4method{getProbTreshold}{ChipDataSet}(object)
}
\arguments{
\item{object}{A \code{\link{ChipDataSet}} object.}
}
\value{
Estimated P(q2) threshold.
}
\description{
Retrieve estimated P(q2) threshold, used to select peaks with a putative
transcription initiation event.
}
\examples{
### Load ChipDataSet object
data(cds)

getProbTreshold(cds)

}
\author{
Armen R. Karapetyan
}
\seealso{
\code{\link{predictStrand}}
}

