% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptionDataSet-generics.R, R/TranscriptionDataSet-methods.R
\docType{methods}
\name{getTestedGapDistances}
\alias{getTestedGapDistances}
\alias{getTestedGapDistances,TranscriptionDataSet-method}
\title{getTestedGapDistances}
\usage{
getTestedGapDistances(object)

\S4method{getTestedGapDistances}{TranscriptionDataSet}(object)
}
\arguments{
\item{object}{A \code{\link{TranscriptionDataSet}} object.}
}
\value{
A \code{data.frame} containing estimated error rates (\code{dissected},
    \code{merged} and \code{sum of two errors}) and corresponding gap distances.
}
\description{
Retrieve a \code{data.frame}, containing a range of tested gap distances
and estimated error rates.
}
\examples{
### Load TranscriptionDataSet object
data(tds)

head(getTestedGapDistances(tds))

}
\author{
Armen R. Karapetyan
}
\seealso{
\code{\link{estimateGapDistance}}
}

