% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transcript.R
\name{read_jhu_transcript}
\alias{read_jhu_transcript}
\alias{read_transcript}
\title{Read JHU Transcript}
\usage{
read_jhu_transcript(file)

read_transcript(file, type = "jhu", ...)
}
\arguments{
\item{file}{path to the filename}

\item{type}{Type of Transcript}

\item{...}{additional arguments to pass to reader function}
}
\value{
A \code{data_frame} of the data
}
\description{
Read JHU Transcript
}
\examples{
pdf_file = "~/Downloads/Transcript.pdf"
if (file.exists(pdf_file)) {
    res = read_jhu_transcript(pdf_file)
}
}
